# MIT-licensed SparseMatrixCSC fkeep! and children