Followup to #13001 and #14631. This pull request replaces the LGPL-licensed SparseMatrixCSC `fkeep!` method and children `tril!`, `triu!`, `droptol!`, and `dropzeros[!]` (noticed signature during cleanup) in base/sparse/csparse.jl with MIT-licensed versions.
    
These methods run in `O(A.n, nnz(A))` time and require no space beyond that passed in. The new methods perform better than the old methods, particularly for large matrices; are the new implementations missing anything that would require additional time?

When editing base/sparse/csparse.jl to remove the existing methods, I folded all blocks to avoid peaking at the LGPL code. So someone should check that the removal was graceful given it was blind but for method signatures.

Benchmarks:
```julia
using Benchmarks: @benchmark, SummaryStatistics

function prettytimes(bench)
    stats = SummaryStatistics(bench)
    timecenter = stats.elapsed_time_center
    timelower = get(stats.elapsed_time_lower)
    timeupper = get(stats.elapsed_time_upper)
    # based on Benchmarks.pretty_time_string
    timecenter < 1_000.0 ? (scalefactor = 1.0; units = "ns") :
        timecenter < 1_000_000.0 ? (scalefactor = 1_000.0; units = "μs") :
            timecenter < 1_000_000_000.0 ? (scalefactor = 1_000_000.0; units = "ms") :
                (scalefactor = 1_000_000_000.0; units = " s")
    @sprintf("%6.2f %s [%6.2f,%6.2f]", timecenter/scalefactor, units, timelower/scalefactor, timeupper/scalefactor)
end

smallN, smallerN = 600, 400;
smallsqrA = sprand(smallN, smallN, 0.01);
smallrectA = sprand(smallN, smallerN, 0.01);
largeN, largerN = 100000, 200000;
largesqrA = sprand(largeN, largeN, 0.001);
largerectA = sprand(largeN, largerN, 0.001);

wm, wt = 12, 25;
println(" $(lpad("method ", wm)) $(lpad("small square A", wt)), $(lpad("small rect A", wt)) | $(lpad("large square A", wt)), $(lpad("large rect A", wt)) ")
@printf("%s: %s , %s | %s, %s\n", lpad("tril!", wm),
    prettytimes(@benchmark triu!(copy(smallsqrA))),
    prettytimes(@benchmark triu!(copy(smallrectA))),
    prettytimes(@benchmark triu!(copy(largesqrA))),
    prettytimes(@benchmark triu!(copy(largerectA))) )
@printf("%s: %s , %s | %s, %s\n", lpad("triu!", wm),
    prettytimes(@benchmark triu!(copy(smallsqrA))),
    prettytimes(@benchmark triu!(copy(smallrectA))),
    prettytimes(@benchmark triu!(copy(largesqrA))),
    prettytimes(@benchmark triu!(copy(largerectA))) )
@printf("%s: %s , %s | %s, %s\n", lpad("droptol!", wm),
    prettytimes(@benchmark Base.SparseArrays.droptol!(copy(smallsqrA), 0.5)),
    prettytimes(@benchmark Base.SparseArrays.droptol!(copy(smallrectA), 0.5)),
    prettytimes(@benchmark Base.SparseArrays.droptol!(copy(largesqrA), 0.5)),
    prettytimes(@benchmark Base.SparseArrays.droptol!(copy(largerectA), 0.5)) )
@printf("%s: %s , %s | %s, %s\n", lpad("dropzeros!", wm),
    prettytimes(@benchmark Base.SparseArrays.dropzeros!(copy(smallsqrA))),
    prettytimes(@benchmark Base.SparseArrays.dropzeros!(copy(smallrectA))),
    prettytimes(@benchmark Base.SparseArrays.dropzeros!(copy(largesqrA))),
    prettytimes(@benchmark Base.SparseArrays.dropzeros!(copy(largerectA))) )
```
On master:
```julia
    method             small square A,              small rect A |            large square A,              large rect A
     tril!:   5.69 μs [  5.62,  5.75] ,   2.50 μs [  2.48,  2.52] |  16.45 ms [ 15.39, 17.51],  47.47 ms [ 45.94, 49.01]
     triu!:   5.66 μs [  5.61,  5.71] ,   2.46 μs [  2.45,  2.47] |  16.32 ms [ 15.33, 17.31],  47.88 ms [ 45.88, 49.89]
  droptol!:   4.64 μs [  4.61,  4.67] ,   3.09 μs [  3.05,  3.13] |  14.32 ms [  9.82, 18.83],  29.77 ms [ 16.43, 43.10]
dropzeros!:  13.10 μs [ 13.07, 13.14] ,   8.57 μs [  8.53,  8.62] |  25.05 ms [ 24.10, 25.99],  50.34 ms [ 49.09, 51.59]
```
On this PR's branch:
```julia
    method             small square A,              small rect A |            large square A,              large rect A
     tril!:   4.31 μs [  4.30,  4.32] ,   1.60 μs [  1.60,  1.61] |   6.51 ms [  5.26,  7.76],  18.19 ms [ 15.06, 21.32]
     triu!:   4.31 μs [  4.29,  4.33] ,   1.60 μs [  1.59,  1.60] |   6.51 ms [  5.27,  7.74],  17.90 ms [ 15.17, 20.64]
  droptol!:   4.47 μs [  4.45,  4.50] ,   2.17 μs [  2.16,  2.19] |   8.59 ms [  3.83, 13.34],  18.41 ms [  4.40, 32.42]
dropzeros!:  10.50 μs [ 10.42, 10.59] ,   6.48 μs [  6.46,  6.51] |  11.45 ms [ 10.88, 12.01],  22.98 ms [ 22.15, 23.81]
```
The `dropzeros!` benchmark isn't particularly meaningful given the absence of stored zeros in the test matrices. Thanks, and best!