@tkelman 2016-01-17
yeah
```
fatal error during bootstrap:
(parse-error "read: unexpected ']'")
ERROR: fatal error loading system image
```
there must be a race condition in a makefile or the parser or something...