@ViralBShah 2016-01-17
Nice performance gain. Would you be able to add some tests as well? Doesn't seem like there are any at the moment.