@Sacha0 2016-01-17
> Would you be able to add some tests as well? Doesn't seem like there are any at the moment.

`tril!` and `triu!` tests exist in 876240c74615761295e419cabf744bd5084f354d/test/sparsedir/sparse.jl#L974-L975. A `droptol!` test exists in 876240c74615761295e419cabf744bd5084f354d/test/sparsedir/sparse.jl#L956-L957. The preceding tests implicitly test `fkeep!`. `dropzeros!` appears to lack tests.

I added a test for `dropzeros!`. Are the existing tests for `tril!`, `triu!`, `droptol!`, and thereby `fkeep!` adequate, or would you like additional tests? If the latter, what would you like to see? Thanks for the review!