@hayd 2016-01-27
@tkelman Are more tests needed here?