@tkelman 2016-01-27
Yes, aside from the clarification about "attach" each of the differences I highlighted means a missing corner case that should be tested against (and would fail until fixed).