@Sacha0 2016-01-27
> each of the differences I highlighted means a missing corner case that should be tested against

Added tests for these edge cases. Thanks!