@tkelman 2016-01-27
ref https://github.com/JuliaLang/julia/blob/91dcdaa6736e2c8287fa3bb3f6f6c7eabd8e5218/base/linalg/dense.jl#L91 for the current bounds check conditions on dense `tril!` - if we're offset this check by one (or two?), we should be consistent between dense and sparse.