@Sacha0 2016-01-27
> we should be consistent between dense and sparse.

Nice catch! Shall fix.