@Sacha0 2016-01-27
Actually, the `k > n || -k > m` test seems inconsistent with the associated error message? An `m`-by-`n` matrix has only `n-1` superdiagonals and `m-1` subdiagonals, so `k == n` and `k == -m` are already out of bounds. Hence the checks I wrote a priori. Am I missing something? If not, perhaps it makes more sense to revise the existing `k > n || -k > m` tests? Thanks!