@tkelman 2016-01-27
Our last few comments have been crossing paths. It does probably make sense to offset the current dense checks by one (though would need to check all the other `tril[!]` and `triu[!]` methods too), but at the very least in a separate commit. And the asymmetry of having one check inclusive and the other exclusive is maybe questionable.