@Sacha0 2016-01-28
> And the asymmetry of having one check inclusive and the other exclusive is maybe questionable.

Agreed. I would be happy to symmetrize the checks. My reasoning for the asymmetric checks was as follows: `tril!(A, A.n - 1)` and `triu!(A, 1 - A.m)` are expensive identities and hence most likely logical errors, whereas `tril!(A, 1 - A.m)` and `triu!(A, A.n - 1)` are legitimate (assuming I did not mix that up).

> Our last few comments have been crossing paths. It does probably make sense to offset the current dense checks by one (though would need to check all the other tril[!] and triu[!] methods too), but at the very least in a separate commit.

Cheers, proposal for action on my part: Keep the offset checks (symmetric, asymmetric, whichever you prefer) in this pull request. Open an issue regarding consistency of `tri(u|l)[!]` `k`-checks. On consensus regarding desired behavior, create a new pull request revising either the methods from this pull request or the other, presently-existing methods in accord with consensus. Thoughts? Thanks again!