@tkelman 2016-01-28
> are expensive identities and hence most likely logical errors

I don't think it's our job to guess what the user's intent was. It may be a silly thing to ask for, but it's valid and within the bounds of the array so I don't think an exception is called for if you ask for the entire array.