@Sacha0 2016-01-28
Checks symmetrized. Thanks!