@tkelman 2016-01-28
32 bit travis failure is a new one, looks unrelated though:
```
julia: /home/travis/build/JuliaLang/julia/src/gc.c:607: find_region: Assertion `maybe && "find_region failed"' failed.

	From worker 3:	     * docs                 Worker 3 terminated.

ERROR (unhandled task failure): EOFError: read end of file
```