@Sacha0 2016-01-28
AppVeyor i686 failures also seem unrelated?