@Sacha0 2016-02-24
Rebased. Is this in shape to merge? Did I miss anything above? Thanks! (Prompted by https://github.com/JuliaLang/julia/pull/14798#discussion_r53892874.)