@tkelman 2016-02-24
This probably doesn't need to use functors any more?