@Sacha0 2016-02-24
> This probably doesn't need to use functors any more?

True, and likewise with `sparse!` and children and `qftranspose!` and children :). If acceptable, I would prefer to merge this pull request and #14798 as they are (with established good performance), and later carefully clean up all those methods in one shot. Does that sound good? Thanks!