@Sacha0 2016-02-25
The Appveyor i686 discontent (failure on 7f845ff/test/arrayops.jl#L1060) seems unrelated?